New Quiz create and display methods
Added create and display functions for Quiz Class. Did not find a way to set and get the openTime and duration yet.
Changed the generateQuizModel algorithm.
Created a main function to test creating and displaying a quiz.
The quiz class works properly without the question bank or the openTime and duration attributes.