Code and Algorithm optimizations.
Implemented a single login function (on the hopes that the ArrayList will store the data correctly.
Removed all login functions on the hopes that a unified login will work.
Updated the DataHandler to have static data structures to access from anywhere instead of passing by reference.
Updated all references of DataHandler members to accomodate for the static changes.
Added interface for console menu function.
Optimized conditions to not need else as much as possible.